pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "lcsgborges",email = "lcsgborges@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.135.3,<0.136.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.9,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "greenlet (>=3.3.2,<4.0.0)",
    "alembic (>=1.18.4,<2.0.0)"
]

[tool.ruff]
line-length = 88
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']
ignore = ['W191']  # TAB

[tool.ruff.format]
preview = true
quote-style = 'single'
indent-style = 'tab'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency= ['thread', 'greenlet']

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=main --cov=core --cov=db --cov=modules -vv'
post_test = 'coverage html'